pyproject.toml
[project]
name = "mcp-server-iceberg"
version = "0.1.0"
description = "MCP server for interacting with an Apache Iceberg catalog and data lake"
readme = "README.md"
requires-python = ">=3.10"  
dependencies = [
    "mcp>=1.0.0",
    "pyiceberg[adlfs,duckdb,pandas]>=0.5.1", # Apache Iceberg Python SDK
    "pyarrow>=14.0.1", # PyArrow for data handling
    "sqlparse>=0.4.4", # SQL parsing
    "python-dotenv", # Environment variable management
    "requests", # HTTP requests for REST catalog
    "fsspec", # Filesystem abstraction
    "s3fs", # S3 filesystem support
]

# Build system configuration 
[build-system]
requires = ["hatchling"]  
build-backend = "hatchling.build" 

# Build configuration 
[tool.hatch.build.targets.wheel]
packages = ["src/mcp_server_iceberg"] 


[project.scripts]
mcp-server-iceberg = "mcp_server_iceberg.server:run_server"  
